requirements.txt
marimo[recommended]~=0.14
rms-vicar~=1.1
pvl~=1.3
numpy~=2.2
matplotlib~=3.10
opencv-python~=4.12
pandas~=2.3
pytest~=8.0
